[FIX] max discount shouldn't prevent sale order copy
Since [1] where a user could not set a discount greated than what he is
allowed to, such users would not be able to copy sale order(s)
containing discount greater than that max allowed discount.

A solution is here provided, to allow the copy by bypassing the max
discount check during the copy, and then reducing the discount which are
greater than the maximum allowed to the maximum allowed discount.

[1]: https://github.com/sbuhl/account_approval/commit/99c5d8bbac65dc6d40bd225f0878a6bc6c131a12

Fixes #7